Split multi-source ingestion step helpers
Move shared BDD context and raw-source helper logic into a focused support
module so the multi-source ingestion step module remains below the line
limit without a file-level lint suppression.

Keep the scenario and step function names in the original discovered test
module, and retain narrow runtime import suppressions required because
`pytest-bdd` inspects step annotations.
